perf(ast): shrink Stmt from 240 to 120 bytes
Two representation changes on the hottest AST types, both costing one
Box or one integer cast:

- `Span` holds `u32` offsets rather than `usize`. A span is embedded in
  nearly every node, so halving it takes `Expr` from 24 bytes to 16, and
  expressions outnumber every other node kind by a wide margin. The
  4 GiB-per-file cap is orders of magnitude above what the parser can
  hold in memory; `Span::new` saturates rather than wrapping.
- `StmtKind::Let` is boxed. `LetStmt` carries an `Ident`, an
  `Option<TypeExpr>` and two `Vec`s, and an enum is as large as its
  largest variant, so inline it forced every assignment, every `wait`,
  every `assert` to the size of a `let`.

  Stmt      240 -> 120     TypeExpr      88 -> 72
  ExprKind  120 ->  96     Ident / Path  40 -> 32
  Expr       24 ->  16

On the bench/split_emit suite (62 MB of source, 352 tests), ABBA-inter-
leaved, 6 runs per arm: peak RSS 3251 -> 2811 MiB, parse 3.5 -> 2.4 s,
wall 13.0 -> 11.9 s. Lowering and emission are unchanged, as expected —
the saving is in the tree itself, and parse-only `harc check` on the
same suite drops 2938 -> 2368 MiB.

Output is unchanged: identical `dump-ir` and identical `harc check`
diagnostics across all 185 fixtures, identical emitted C++ across 330
fixture/codegen combinations, and the 348 MB split output byte-for-byte
unchanged.

+++ src/codegen/cpp_tb.rs
@@ -5869,7 +5869,7 @@ struct Emitter {
     /// Runtime problem id per concrete `randomize(...)` target span.
     /// Emitted as metadata touches only during the scaffold phase; current
     /// behavior still uses the existing PRNG / inline Z3 solve paths.
-    runtime_randomize_problem_ids: std::collections::HashMap<(usize, usize), u32>,
+    runtime_randomize_problem_ids: std::collections::HashMap<(u32, u32), u32>,
     /// Free-standing `relation` declarations indexed by name (spec
     /// §4.2). `Call(Ident(R), args)` inside a constraint expression
     /// expands to R's body with formal parameters substituted by the
@@ -5958,7 +5958,7 @@ struct Emitter {
     /// In-flight substitutions for temporal SystemCall expressions during
     /// property emission. Keyed by AST span so the rewrite finds the right
     /// occurrence. Set/cleared around each `emit_property_check` call.
-    prop_subs: std::collections::HashMap<(usize, usize), String>,
+    prop_subs: std::collections::HashMap<(u32, u32), String>,
     /// Event name → C++ inner type (e.g. `uint64_t`). Populated when
     /// emitting `let e : event<T>`; consulted when emitting `emit e(arg)`
     /// and `on e(arg) ... end on` so the lambda gets the right param type.
